migrate set up cubes method to recursive filter
